jmoney-backend/src/main/java/dev/kociolek/jmoney/service/UserService.java
package dev.kociolek.jmoney.service;

import dev.kociolek.jmoney.domain.User;
import dev.kociolek.jmoney.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserService implements IService<User> {

	private UserRepository repository;

	@Autowired
	public UserService(UserRepository repository) {
		this.repository = repository;
	}

	@Override
	public User create(User user) {
		return repository.save(user);
	}

	@Override
	public User update(User user) {
		return repository.save(user);
	}

	@Override
	public void delete(User user) {
		repository.delete(user);
	}

	@Override
	public Optional<User> findById(String id) {
		return repository.findById(id);
	}


}